refactor: 모듈 구조 재정비 및 코드 리팩토링
- 기존 모듈 구조에서 불필요한 모듈 병합 및 제거 처리
- 헥사고날 아키텍처를 기반으로 하되 생산성을 저해하지 않도록 재정비
- `adapter`를 `infrastructure`에 포함시켜 구조 단순화
- 구조 리팩토링 과정에서 대량의 파일 이동, 추가 및 삭제 발생
- 이에 따른 package 및 import 문 수정
- 변경된 모듈 구조 변경을 반영해 README 파일 수정
