--- movie-adapter/build.gradle
@@ -1,10 +1,0 @@
-bootJar { enabled = false }
-jar { enabled = true }
-
-dependencies {
-    implementation project(':movie-application') // port 사용 위해 추가
-    implementation project(':movie-domain')
-    implementation project(':movie-common')
-    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
-    runtimeOnly 'com.mysql:mysql-connector-j' // DB 드라이버
-}
