--- movie-adapter/src/main/java/com/example/movieadapter/db/MovieJpaRepository.java
@@ -1,7 +1,0 @@
-package com.example.movieadapter.db;
-
-import com.example.moviedomain.entity.Movie;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MovieJpaRepository extends JpaRepository<Movie, Long> {
-}
