--- movie-api/build.gradle
@@ -1,10 +1,0 @@
-bootJar { enabled = true }
-jar { enabled = false }
-
-dependencies {
-	implementation project(':movie-application')
-	implementation project(':movie-common')
-	implementation project(':movie-adapter') // MovieApiRepositoryPort 구현체 찾을 수 없어서 의존성 등록
-	implementation 'org.springframework.boot:spring-boot-starter-web'
-	implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
-}
