+++ movie-application/build.gradle
@@ -3,5 +3,7 @@ jar { enabled = true }
 
 dependencies {
 	implementation project(':movie-domain')
-	implementation project(':movie-common')
+	implementation("org.springframework.boot:spring-boot-starter-web")
+	implementation ("org.springframework.boot:spring-boot-starter-data-jpa")
+	implementation("org.springframework.boot:spring-boot-starter-validation")
 }
