+++ movie-application/src/main/java/com/example/application/dto/request/ScreeningRequestDto.java
@@ -1,0 +1,17 @@
+package com.example.application.dto.request;
+
+import jakarta.validation.constraints.NotNull;
+import java.time.LocalTime;
+
+public record ScreeningRequestDto(
+
+        @NotNull(message = "Movie must not be null")
+        Long movieId,
+
+        @NotNull(message = "Theater must not be null")
+        Long theaterId,
+
+        @NotNull(message = "Start time must not be null")
+        LocalTime startTime
+
+        ) {}
