+++ movie-application/src/main/java/com/example/application/port/in/MovieServicePort.java
@@ -1,0 +1,10 @@
+package com.example.application.port.in;
+
+import com.example.application.dto.request.MovieRequestDto;
+import com.example.application.dto.response.MovieResponseDto;
+import java.util.List;
+
+public interface MovieServicePort {
+    List<MovieResponseDto> getMoviesSortedByReleaseDateDesc();
+    void createMovie(MovieRequestDto movieRequestDto);
+}
