--- movie-application/src/main/java/com/example/movieapplication/port/in/MovieServicePort.java
@@ -1,8 +1,0 @@
-package com.example.movieapplication.port.in;
-
-import com.example.movieapplication.dto.MovieResponseDto;
-import java.util.List;
-
-public interface MovieServicePort {
-    List<MovieResponseDto> getAllMovies();
-}
