--- movie-common/build.gradle
@@ -1,6 +1,0 @@
-bootJar { enabled = false }
-jar { enabled = true }
-
-dependencies {
-    implementation 'org.springframework.boot:spring-boot-starter-web'
-}
