--- movie-common/src/main/java/com/example/moviecommon/exception/CustomException.java
@@ -1,19 +1,0 @@
-package com.example.moviecommon.exception;
-
-import lombok.Getter;
-
-@Getter
-public class CustomException extends RuntimeException {
-    private final ErrorCode errorCode;
-
-    public CustomException(ErrorCode errorCode) {
-        super(errorCode.getMessage());
-        this.errorCode = errorCode;
-    }
-
-    public CustomException(ErrorCode errorCode, String customMessage) {
-        super(customMessage);
-        this.errorCode = errorCode;
-    }
-
-}
