--- movie-common/src/main/java/com/example/moviecommon/exception/ErrorResponse.java
@@ -1,4 +1,0 @@
-package com.example.moviecommon.exception;
-
-public record ErrorResponse(String errorCode, String message, int statusCode) {
-}
