--- movie-common/src/main/java/com/example/moviecommon/exception/GlobalExceptionHandler.java
@@ -1,28 +1,0 @@
-package com.example.moviecommon.exception;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-@RestControllerAdvice
-public class GlobalExceptionHandler {
-    @ExceptionHandler(CustomException.class)
-    public ResponseEntity<ErrorResponse> handleCustomException(CustomException ex) {
-        ErrorResponse errorResponse = new ErrorResponse(
-                ex.getErrorCode().name(),
-                ex.getErrorCode().getMessage(),
-                ex.getErrorCode().getStatusCode()
-        );
-        return ResponseEntity.status(ex.getErrorCode().getStatusCode()).body(errorResponse);
-    }
-
-    @ExceptionHandler(Exception.class)
-    public ResponseEntity<ErrorResponse> handleGeneralException(Exception ex) {
-        ErrorResponse errorResponse = new ErrorResponse(
-                "INTERNAL_SERVER_ERROR",
-                ex.getMessage(),
-                500
-        );
-        return ResponseEntity.status(500).body(errorResponse);
-    }
-}
