+++ movie-domain/build.gradle
@@ -2,6 +2,5 @@ bootJar { enabled = false }
 jar { enabled = true }
 
 dependencies {
-    implementation project(':movie-common')
     implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
 }
