+++ movie-domain/src/main/java/com/example/domain/exception/ErrorResponse.java
@@ -1,0 +1,4 @@
+package com.example.domain.exception;
+
+public record ErrorResponse(String errorCode, String message, int statusCode) {
+}
