--- movie-domain/src/main/java/com/example/moviedomain/base/AuditingFields.java
@@ -1,36 +1,0 @@
-package com.example.moviedomain.base;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.EntityListeners;
-import jakarta.persistence.MappedSuperclass;
-import java.time.LocalDateTime;
-import org.springframework.data.annotation.CreatedBy;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedBy;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-import org.springframework.format.annotation.DateTimeFormat;
-
-@MappedSuperclass
-@EntityListeners(AuditingEntityListener.class)
-public abstract class AuditingFields {
-
-    @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME)
-    @CreatedDate
-    @Column(nullable = false, updatable = false)
-    protected LocalDateTime createdAt;
-
-    @CreatedBy
-    @Column(nullable = false, updatable = false)
-    protected String createdBy;
-
-    @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME)
-    @LastModifiedDate
-    @Column(nullable = false)
-    protected LocalDateTime modifiedAt;
-
-    @LastModifiedBy
-    @Column(nullable = false)
-    protected String modifiedBy;
-
-}
