+++ movie-infrastructure/build.gradle
@@ -1,0 +1,11 @@
+bootJar { enabled = false }
+jar { enabled = true }
+
+dependencies {
+    implementation project(':movie-application')
+    implementation project(':movie-domain')
+    implementation 'org.springframework.boot:spring-boot-starter-web'
+    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
+    implementation("org.springframework.boot:spring-boot-starter-validation")
+    runtimeOnly 'com.mysql:mysql-connector-j' // DB 드라이버
+}
