+++ movie-infrastructure/src/main/java/com/example/infrastructure/db/SeatJpaRepository.java
@@ -1,0 +1,7 @@
+package com.example.infrastructure.db;
+
+import com.example.domain.model.entity.Seat;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface SeatJpaRepository extends JpaRepository<Seat, Long> {
+}
