+++ movie-infrastructure/src/main/java/com/example/infrastructure/db/TheaterJpaRepository.java
@@ -1,0 +1,7 @@
+package com.example.infrastructure.db;
+
+import com.example.domain.model.entity.Theater;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface TheaterJpaRepository extends JpaRepository<Theater, Long> {
+}
